ai-service/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
flwr>=1.4.0
syft>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
redis>=4.6.0
celery>=5.3.0
ipfshttpclient>=0.8.0
web3>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tree-sitter>=0.20.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0 